more fixes
- update to warn when scraper is missing a "see more" value in some text to help identify when project is missing button clicks.
- Fixed logger reporting cleanmessagedata as filename inside cleanprofiledata file.
- updated position description cleanup to remove "see more" from the final value as well as "see less"
- fix to typo in template update made yesterday for Company name selector.
- fixed grammer in scrapedin warning messages (modifed past tense "was" with "were")
- updated see more buttons to expan recommendations and fixed selectors for summary alternative and positions.

TODO Seemorebuttons.js needs a modification because when a user has many recommendations the clicks between "received" and "given" cause the html to roll up previous expanded results. The correct way to collect all of the data will be to expand "received" and collect the data and THEN go to "given" - expand the section and collect the data.